+++ example/src/main/java/org/wordpress/android/fluxc/example/MainExampleActivity.java
@@ -1,505 +1,30 @@
 package org.wordpress.android.fluxc.example;
 
 import android.app.Activity;
-import android.app.AlertDialog;
-import android.app.DialogFragment;
-import android.app.FragmentTransaction;
-import android.content.DialogInterface;
 import android.os.Bundle;
-import android.text.TextUtils;
-import android.view.View;
-import android.view.View.OnClickListener;
-import android.widget.Button;
-import android.widget.EditText;
+import android.text.method.ScrollingMovementMethod;
 import android.widget.TextView;
-
-import org.greenrobot.eventbus.Subscribe;
-import org.greenrobot.eventbus.ThreadMode;
-import org.wordpress.android.fluxc.Dispatcher;
-import org.wordpress.android.fluxc.action.PostAction;
-import org.wordpress.android.fluxc.example.ThreeEditTextDialog.Listener;
-import org.wordpress.android.fluxc.generated.AccountActionBuilder;
-import org.wordpress.android.fluxc.generated.AuthenticationActionBuilder;
-import org.wordpress.android.fluxc.generated.PostActionBuilder;
-import org.wordpress.android.fluxc.generated.SiteActionBuilder;
-import org.wordpress.android.fluxc.model.PostModel;
-import org.wordpress.android.fluxc.model.SiteModel;
-import org.wordpress.android.fluxc.network.HTTPAuthManager;
-import org.wordpress.android.fluxc.network.MemorizingTrustManager;
-import org.wordpress.android.fluxc.network.discovery.SelfHostedEndpointFinder.DiscoveryError;
-import org.wordpress.android.fluxc.store.AccountStore;
-import org.wordpress.android.fluxc.store.AccountStore.AuthenticatePayload;
-import org.wordpress.android.fluxc.store.AccountStore.NewAccountPayload;
-import org.wordpress.android.fluxc.store.AccountStore.OnAccountChanged;
-import org.wordpress.android.fluxc.store.AccountStore.OnAuthenticationChanged;
-import org.wordpress.android.fluxc.store.AccountStore.OnDiscoveryResponse;
-import org.wordpress.android.fluxc.store.AccountStore.OnNewUserCreated;
-import org.wordpress.android.fluxc.store.AccountStore.PushAccountSettingsPayload;
-import org.wordpress.android.fluxc.store.PostStore;
-import org.wordpress.android.fluxc.store.PostStore.FetchPostsPayload;
-import org.wordpress.android.fluxc.store.PostStore.InstantiatePostPayload;
-import org.wordpress.android.fluxc.store.PostStore.OnPostChanged;
-import org.wordpress.android.fluxc.store.PostStore.OnPostInstantiated;
-import org.wordpress.android.fluxc.store.PostStore.OnPostUploaded;
-import org.wordpress.android.fluxc.store.PostStore.RemotePostPayload;
-import org.wordpress.android.fluxc.store.SiteStore;
-import org.wordpress.android.fluxc.store.SiteStore.OnSiteChanged;
-import org.wordpress.android.fluxc.store.SiteStore.OnSiteRemoved;
-import org.wordpress.android.fluxc.store.SiteStore.RefreshSitesXMLRPCPayload;
-import org.wordpress.android.util.AppLog;
-import org.wordpress.android.util.AppLog.T;
-import org.wordpress.android.util.ToastUtils;
-
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.HashMap;
-import java.util.List;
-
-import javax.inject.Inject;
 
 
 public class MainExampleActivity extends Activity {
-    @Inject SiteStore mSiteStore;
-    @Inject AccountStore mAccountStore;
-    @Inject PostStore mPostStore;
-    @Inject Dispatcher mDispatcher;
-    @Inject HTTPAuthManager mHTTPAuthManager;
-    @Inject MemorizingTrustManager mMemorizingTrustManager;
-
     private TextView mLogView;
-
-    private Button mCreatePostOnFirstSite;
-    private Button mDeleteLatestPost;
-    private Button mAccountInfos;
-    private Button mAccountSettings;
-    private Button mFetchFirstSitePosts;
-
-    // Would be great to not have to keep this state, but it makes HTTPAuth and self signed SSL management easier
-    private RefreshSitesXMLRPCPayload mSelfhostedPayload;
-
-    // used for 2fa
-    private AuthenticatePayload mAuthenticatePayload;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         ((ExampleApp) getApplication()).component().inject(this);
         setContentView(R.layout.activity_example);
-        mAccountInfos = (Button) findViewById(R.id.account_infos_button);
-        mAccountInfos.setOnClickListener(new OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                mDispatcher.dispatch(AccountActionBuilder.newFetchAccountAction());
-                mDispatcher.dispatch(AccountActionBuilder.newFetchSettingsAction());
-            }
-        });
 
-        Button signIn = (Button) findViewById(R.id.sign_in_fetch_sites_button);
-        signIn.setOnClickListener(new OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                // show signin dialog
-                showSigninDialog();
-            }
-        });
-        mCreatePostOnFirstSite = (Button) findViewById(R.id.create_new_post_first_site);
-        mCreatePostOnFirstSite.setOnClickListener(new OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                PostStore.InstantiatePostPayload payload = new InstantiatePostPayload(mSiteStore.getSites().get(0),
-                        false);
-                mDispatcher.dispatch(PostActionBuilder.newInstantiatePostAction(payload));
-            }
-        });
-        mFetchFirstSitePosts = (Button) findViewById(R.id.fetch_first_site_posts);
-        mFetchFirstSitePosts.setOnClickListener(new OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                FetchPostsPayload payload = new FetchPostsPayload(mSiteStore.getSites().get(0));
-                mDispatcher.dispatch(PostActionBuilder.newFetchPostsAction(payload));
-            }
-        });
-        mDeleteLatestPost = (Button) findViewById(R.id.delete_a_post_from_first_site);
-        mDeleteLatestPost.setOnClickListener(new OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                SiteModel firstSite = mSiteStore.getSites().get(0);
-                List<PostModel> posts = mPostStore.getPostsForSite(firstSite);
-                Collections.sort(posts, new Comparator<PostModel>() {
-                    @Override
-                    public int compare(PostModel lhs, PostModel rhs) {
-                        return (int) (rhs.getRemotePostId() - lhs.getRemotePostId());
-                    }
-                });
-                if (!posts.isEmpty()) {
-                    RemotePostPayload payload = new RemotePostPayload(posts.get(0), firstSite);
-                    mDispatcher.dispatch(PostActionBuilder.newDeletePostAction(payload));
-                }
-            }
-        });
-
-        mAccountSettings = (Button) findViewById(R.id.account_settings);
-        mAccountSettings.setOnClickListener(new OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                changeAccountSettings();
-            }
-        });
-
-        Button signOut = (Button) findViewById(R.id.signout);
-        signOut.setOnClickListener(new OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                signOut();
-            }
-        });
-
-        Button newAccount = (Button) findViewById(R.id.new_account);
-        newAccount.setOnClickListener(new OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                showNewAccountDialog();
-            }
-        });
-
-        Button sites = (Button) findViewById(R.id.sites);
-        sites.setOnClickListener(new OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                if (mSiteStore.getSitesCount() == 0 && !mAccountStore.hasAccessToken()) {
-                    ToastUtils.showToast(MainExampleActivity.this, "You must be logged in");
-                    return;
-                }
-                SitesFragment s = new SitesFragment();
-                getFragmentManager().beginTransaction()
-                        .replace(R.id.actions_fragment, s)
-                        .commit();
-            }
-        });
+        MainFragment mf = new MainFragment();
+        getFragmentManager().beginTransaction().add(R.id.fragment_container, mf).commit();
 
         mLogView = (TextView) findViewById(R.id.log);
-
-        init();
-    }
-
-    private void init() {
-        mAccountInfos.setEnabled(mAccountStore.hasAccessToken());
-        mAccountSettings.setEnabled(mAccountStore.hasAccessToken());
-        if (mAccountStore.hasAccessToken()) {
-            prependToLog("You're signed in as: " + mAccountStore.getAccount().getUserName());
-        }
-        // FIXME: mSites.setEnabled(mSiteStore.hasSite());
-        mCreatePostOnFirstSite.setEnabled(mSiteStore.hasSite());
-        mDeleteLatestPost.setEnabled(mSiteStore.hasSite());
-    }
-
-    @Override
-    protected void onStart() {
-        super.onStart();
-        // Order is important here since onRegister could fire onChanged events. "register(this)" should probably go
-        // first everywhere.
-        mDispatcher.register(this);
-    }
-
-    @Override
-    protected void onStop() {
-        super.onStop();
-        mDispatcher.unregister(this);
+        mLogView.setMovementMethod(new ScrollingMovementMethod());
+        prependToLog("I'll log stuff here.");
     }
 
     public void prependToLog(final String s) {
         String output = s + "\n" + mLogView.getText();
         mLogView.setText(output);
-    }
-
-
-    // Private methods
-
-    private void showSSLWarningDialog(String certifString) {
-        FragmentTransaction ft = getFragmentManager().beginTransaction();
-        SSLWarningDialog newFragment = SSLWarningDialog.newInstance(
-                new android.content.DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which) {
-                        // Add the certificate to our list
-                        mMemorizingTrustManager.storeLastFailure();
-                        // Retry login action
-                        if (mSelfhostedPayload != null) {
-                            signInAction(mSelfhostedPayload.username, mSelfhostedPayload.password,
-                                    mSelfhostedPayload.url);
-                        }
-                    }
-                }, certifString);
-        newFragment.show(ft, "dialog");
-    }
-
-    private void showSigninDialog() {
-        FragmentTransaction ft = getFragmentManager().beginTransaction();
-        DialogFragment newFragment = ThreeEditTextDialog.newInstance(new Listener() {
-            @Override
-            public void onClick(String username, String password, String url) {
-                signInAction(username, password, url);
-            }
-        }, "Username", "Password", "XMLRPC Url");
-        newFragment.show(ft, "dialog");
-    }
-
-    private void show2faDialog() {
-        FragmentTransaction ft = getFragmentManager().beginTransaction();
-        DialogFragment newFragment = ThreeEditTextDialog.newInstance(new Listener() {
-            @Override
-            public void onClick(String text1, String text2, String text3) {
-                if (TextUtils.isEmpty(text3)) {
-                    prependToLog("2FA code required to login");
-                    return;
-                }
-                signIn2fa(text3);
-            }
-        }, "", "", "2FA Code");
-        newFragment.show(ft, "2fadialog");
-    }
-
-    private void signIn2fa(String twoStepCode) {
-        mAuthenticatePayload.twoStepCode = twoStepCode;
-        mAuthenticatePayload.nextAction = SiteActionBuilder.newFetchSitesAction();
-        mDispatcher.dispatch(AuthenticationActionBuilder.newAuthenticateAction(mAuthenticatePayload));
-    }
-
-    private void showHTTPAuthDialog(final String url) {
-        FragmentTransaction ft = getFragmentManager().beginTransaction();
-        DialogFragment newFragment = ThreeEditTextDialog.newInstance(new Listener() {
-            @Override
-            public void onClick(String username, String password, String unused) {
-                // Add credentials
-                mHTTPAuthManager.addHTTPAuthCredentials(username, password, url, null);
-                // Retry login action
-                if (mSelfhostedPayload != null) {
-                    signInAction(mSelfhostedPayload.username, mSelfhostedPayload.password, url);
-                }
-            }
-        }, "Username", "Password", "unused");
-        newFragment.show(ft, "dialog");
-    }
-
-    /**
-     * Called when the user tap OK on the SignIn Dialog. It authenticates and list user sites, on wpcom or self hosted
-     * depending if the user filled the URL field.
-     */
-    private void signInAction(final String username, final String password, final String url) {
-        if (TextUtils.isEmpty(url)) {
-            wpcomFetchSites(username, password);
-        } else {
-            mSelfhostedPayload = new RefreshSitesXMLRPCPayload();
-            mSelfhostedPayload.url = url;
-            mSelfhostedPayload.username = username;
-            mSelfhostedPayload.password = password;
-
-            mDispatcher.dispatch(AuthenticationActionBuilder.newDiscoverEndpointAction(mSelfhostedPayload));
-        }
-    }
-
-    private void signOut() {
-        mDispatcher.dispatch(AccountActionBuilder.newSignOutAction());
-        mDispatcher.dispatch(SiteActionBuilder.newRemoveWpcomSitesAction());
-        // Remove all remaining sites
-        for (SiteModel site : mSiteStore.getSites()) {
-            mDispatcher.dispatch(SiteActionBuilder.newRemoveSiteAction(site));
-        }
-    }
-
-    private void wpcomFetchSites(String username, String password) {
-        mAuthenticatePayload = new AuthenticatePayload(username, password);
-        // Next action will be dispatched if authentication is successful
-        mAuthenticatePayload.nextAction = SiteActionBuilder.newFetchSitesAction();
-        mDispatcher.dispatch(AuthenticationActionBuilder.newAuthenticateAction(mAuthenticatePayload));
-    }
-
-    private void selfHostedFetchSites(String username, String password, String xmlrpcEndpoint) {
-        RefreshSitesXMLRPCPayload payload = new RefreshSitesXMLRPCPayload();
-        payload.username = username;
-        payload.password = password;
-        payload.url = xmlrpcEndpoint;
-        mSelfhostedPayload = payload;
-        // Self Hosted don't have any "Authentication" request, try to list sites with user/password
-        mDispatcher.dispatch(SiteActionBuilder.newFetchSitesXmlRpcAction(payload));
-    }
-
-    private void changeAccountSettings() {
-        AlertDialog.Builder alert = new AlertDialog.Builder(this);
-        final EditText edittext = new EditText(this);
-        alert.setMessage("Update your display name:");
-        alert.setView(edittext);
-        alert.setPositiveButton(android.R.string.ok, new DialogInterface.OnClickListener() {
-            public void onClick(DialogInterface dialog, int whichButton) {
-                String displayName = edittext.getText().toString();
-                PushAccountSettingsPayload payload = new PushAccountSettingsPayload();
-                payload.params = new HashMap<>();
-                payload.params.put("display_name", displayName);
-                mDispatcher.dispatch(AccountActionBuilder.newPushSettingsAction(payload));
-            }
-        });
-        alert.show();
-    }
-
-    private void showNewAccountDialog() {
-        FragmentTransaction ft = getFragmentManager().beginTransaction();
-        DialogFragment newFragment = ThreeEditTextDialog.newInstance(new Listener() {
-            @Override
-            public void onClick(String username, String email, String password) {
-                newAccountAction(username, email, password);
-            }
-        }, "Username", "Email", "Password");
-        newFragment.show(ft, "dialog");
-    }
-
-    private void newAccountAction(String username, String email, String password) {
-        NewAccountPayload newAccountPayload = new NewAccountPayload(username, password, email, true);
-        mDispatcher.dispatch(AccountActionBuilder.newCreateNewAccountAction(newAccountPayload));
-    }
-
-    // Event listeners
-
-    @Subscribe(threadMode = ThreadMode.MAIN)
-    public void onAccountChanged(OnAccountChanged event) {
-        if (!mAccountStore.hasAccessToken()) {
-            prependToLog("Signed Out");
-        } else {
-            if (event.accountInfosChanged) {
-                prependToLog("Display Name: " + mAccountStore.getAccount().getDisplayName());
-            }
-        }
-    }
-
-    @Subscribe(threadMode = ThreadMode.MAIN)
-    public void onAuthenticationChanged(OnAuthenticationChanged event) {
-        mAccountInfos.setEnabled(mAccountStore.hasAccessToken());
-        mAccountSettings.setEnabled(mAccountStore.hasAccessToken());
-        if (event.isError()) {
-            prependToLog("Authentication error: " + event.error.type);
-
-            switch (event.error.type) {
-                case AUTHORIZATION_REQUIRED:
-                case ACCESS_DENIED:
-                    // You're not authorized to do that
-                    break;
-                case INVALID_CLIENT:
-                case INVALID_GRANT:
-                case UNSUPPORTED_GRANT_TYPE:
-                    // You should fix your gradle.properties
-                    break;
-                case UNKNOWN_TOKEN:
-                case INVALID_TOKEN:
-                case NOT_AUTHENTICATED:
-                case INCORRECT_USERNAME_OR_PASSWORD:
-                    showSigninDialog();
-                    break;
-                case HTTP_AUTH_ERROR:
-                    // Show a Dialog prompting for http username and password
-                    showHTTPAuthDialog(mSelfhostedPayload.url);
-                    break;
-                case INVALID_SSL_CERTIFICATE:
-                    // Show a SSL Warning Dialog
-                    showSSLWarningDialog(mMemorizingTrustManager.getLastFailure().toString());
-                    break;
-                case NEEDS_2FA:
-                    show2faDialog();
-                    break;
-                case INVALID_OTP:
-                    break;
-                case INVALID_REQUEST:
-                case UNSUPPORTED_RESPONSE_TYPE:
-                case GENERIC_ERROR:
-                    // Show Toast "Network Error"?
-                    break;
-            }
-        }
-    }
-
-    @Subscribe(threadMode = ThreadMode.MAIN)
-    public void onDiscoveryResponse(OnDiscoveryResponse event) {
-        if (event.isError()) {
-            if (event.error == DiscoveryError.WORDPRESS_COM_SITE) {
-                wpcomFetchSites(mSelfhostedPayload.username, mSelfhostedPayload.password);
-            } else if (event.error == DiscoveryError.HTTP_AUTH_REQUIRED) {
-                showHTTPAuthDialog(event.failedEndpoint);
-            } else if (event.error == DiscoveryError.ERRONEOUS_SSL_CERTIFICATE) {
-                mSelfhostedPayload.url = event.failedEndpoint;
-                showSSLWarningDialog(mMemorizingTrustManager.getLastFailure().toString());
-            }
-            prependToLog("Discovery failed with error: " + event.error);
-            AppLog.e(T.API, "Discover error: " + event.error);
-        } else {
-            prependToLog("Discovery succeeded, endpoint: " + event.xmlRpcEndpoint);
-            selfHostedFetchSites(mSelfhostedPayload.username, mSelfhostedPayload.password, event.xmlRpcEndpoint);
-        }
-    }
-
-    @Subscribe(threadMode = ThreadMode.MAIN)
-    public void onSiteChanged(OnSiteChanged event) {
-        if (event.isError()) {
-            prependToLog("SiteChanged error: " + event.error.type);
-            return;
-        }
-        if (mSiteStore.hasSite()) {
-            SiteModel firstSite = mSiteStore.getSites().get(0);
-            prependToLog("First site name: " + firstSite.getName() + " - Total sites: " + mSiteStore.getSitesCount());
-            // FIXME: mSites.setEnabled(true);
-        } else {
-            // FIXME: mSites..setEnabled(false);
-        }
-    }
-
-    @Subscribe(threadMode = ThreadMode.MAIN)
-    public void onNewUserValidated(OnNewUserCreated event) {
-        String message = event.dryRun ? "validation" : "creation";
-        if (event.isError()) {
-            prependToLog("New user " + message + ": error: " + event.error.type + " - " + event.error.message);
-        } else {
-            prependToLog("New user " + message + ": success!");
-        }
-    }
-
-    @Subscribe(threadMode = ThreadMode.MAIN)
-    public void onSiteRemoved(OnSiteRemoved event) {
-        // FIXME: mSites.setEnabled(mSiteStore.hasSite());
-    }
-
-    @Subscribe(threadMode = ThreadMode.MAIN)
-    public void onPostChanged(OnPostChanged event) {
-        if (event.isError()) {
-            prependToLog("Error from " + event.causeOfChange + " - error: " + event.error.type);
-            return;
-        }
-
-        SiteModel firstSite = mSiteStore.getSites().get(0);
-        if (!mPostStore.getPostsForSite(firstSite).isEmpty()) {
-            if (event.causeOfChange.equals(PostAction.FETCH_POSTS)
-                || event.causeOfChange.equals(PostAction.FETCH_PAGES)) {
-                prependToLog("Fetched " + event.rowsAffected + " posts from: " + firstSite.getName());
-            } else if (event.causeOfChange.equals(PostAction.DELETE_POST)) {
-                prependToLog("Post deleted!");
-            }
-            mCreatePostOnFirstSite.setEnabled(true);
-            mDeleteLatestPost.setEnabled(true);
-        } else {
-            mCreatePostOnFirstSite.setEnabled(false);
-            mDeleteLatestPost.setEnabled(false);
-        }
-    }
-
-    @Subscribe(threadMode = ThreadMode.MAIN)
-    public void onPostInstantiated(OnPostInstantiated event) {
-        PostModel examplePost = event.post;
-        examplePost.setTitle("From example activity");
-        examplePost.setContent("Hi there, I'm a post from FluxC!");
-        examplePost.setFeaturedImageId(0);
-
-        RemotePostPayload payload = new RemotePostPayload(examplePost, mSiteStore.getSites().get(0));
-        mDispatcher.dispatch(PostActionBuilder.newPushPostAction(payload));
-    }
-
-    @Subscribe(threadMode = ThreadMode.MAIN)
-    public void onPostUploaded(OnPostUploaded event) {
-        prependToLog("Post uploaded! Remote post id: " + event.post.getRemotePostId());
     }
 }
