+++ example/src/main/res/layout/fragment_account.xml
@@ -1,0 +1,9 @@
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+              xmlns:tools="http://schemas.android.com/tools"
+              android:layout_width="match_parent"
+              android:layout_height="match_parent"
+              android:orientation="vertical"
+              tools:context="org.wordpress.android.fluxc.example.AccountFragment">
+
+
+</LinearLayout>
