+++ example/src/main/res/layout/fragment_main.xml
@@ -66,38 +66,9 @@
         android:text="Comments" />
 
     <Button
-        android:id="@+id/account_settings"
+        android:id="@+id/media"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:enabled="false"
-        android:text="Set Display Name" />
-
-    <Button
-        android:id="@+id/account_infos_button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:enabled="false"
-        android:text="Fetch Account" />
-
-    <Button
-        android:id="@+id/fetch_first_site_posts"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:enabled="false"
-        android:text="Fetch posts from first site" />
-
-    <Button
-        android:id="@+id/create_new_post_first_site"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:enabled="false"
-        android:text="Create a new post on first site" />
-
-    <Button
-        android:id="@+id/delete_a_post_from_first_site"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:enabled="false"
-        android:text="Delete a post from first site" />
+        android:text="Media" />
 
 </LinearLayout>
